requirements_unittests_venv.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.2.post1
argcomplete==3.5.1
attrs==24.2.0
azure-core==1.31.0
azure-monitor-opentelemetry-exporter==1.0.0b31
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
dataclasses-json==0.6.7
datasets==3.0.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
fastapi==0.115.2
filelock==3.16.1
filetype==1.2.0
fixedint==0.1.6
Flask==3.0.3
Flask-Cors==4.0.2
flask-restx==1.3.0
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google_search_results==2.4.1
googleapis-common-protos==1.65.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
iniconfig==2.0.0
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.4
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==24.3.1
langchain==0.3.4
langchain-community==0.3.3
langchain-core==0.3.12
langchain-openai==0.2.3
langchain-text-splitters==0.3.0
langsmith==0.1.136
Levenshtein==0.25.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
mdurl==0.1.2
more-itertools==10.5.0
msrest==0.7.1
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.52.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opik==0.2.2
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pluggy==1.5.0
prompt-toolkit==3.0.36
promptflow==1.16.1
promptflow-core==1.16.1
promptflow-devkit==1.16.1
promptflow-tools==1.4.0
promptflow-tracing==1.16.1
propcache==0.2.0
protobuf==4.25.5
psutil==5.9.8
psycopg2-binary==2.9.10
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydash==7.0.7
Pygments==2.18.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
questionary==2.0.1
RapidFuzz==3.10.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
strictyaml==1.7.3
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uuid7==0.1.0
waitress==2.1.2
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
xxhash==3.5.0
yarl==1.15.5
zipp==3.20.2
